Refactor some scripts for the UI change that will be made
- Added a new scene, without a dual camera set up.
- Some latent meta files were committed.
