+++ Entelect Challenge 2019 Visualiser/Assets/Clean Vector Icons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1bfc5d682a4755848a447fffefda67f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
