+++ Entelect Challenge 2019 Visualiser/Assets/Clean Vector Icons/Update.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9fea51cdf55af8f42a336882b1f9d75c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
