+++ Entelect Challenge 2019 Visualiser/Assets/Essential_Isometric_3D_Block_Pack_Devilsworkshop_v2.0.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b1568a224f93995489115b61e21a780c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
