+++ Entelect Challenge 2019 Visualiser/Assets/Essential_Isometric_3D_Block_Pack_Devilsworkshop_v2.0/Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 18274f4aaa9661c4d8e4f0cd4f126bde
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
