+++ Entelect Challenge 2019 Visualiser/Assets/Essential_Isometric_3D_Block_Pack_Devilsworkshop_v2.0/Assets/OBJ.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f992d5b3a0e4324d88c13a62af43382
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
