+++ Entelect Challenge 2019 Visualiser/Assets/Spritesheets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58f99c3439327124da739ae640220916
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
