+++ Entelect Challenge 2019 Visualiser/Assets/Spritesheets/explosions-pack-web.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7cb108036289586449e050d58aacded8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
