+++ Entelect Challenge 2019 Visualiser/Assets/Spritesheets/explosions-pack-web/spritesheets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db5812972c5247441ba8672c760d8bc7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
